Fix tests that were broken due to refactoring reducks architecture
